requirements.txt
flask==2.0.1
flask_sqlalchemy==2.5.1
sqlalchemy==1.3.8
bcrypt==4.0.1
google-cloud-bigquery==3.11.1
flask_apscheduler==1.12.4
pytest==7.4.0
psycopg2==2.9.6